backend/requirements.txt
Pillow
numpy
requests
tensorflow